requirements.txt
streamlit_webrtc
genai
ibm-generative-ai
streamlit
google-generativeai